Refactor for testing, add minimal unit tests. Add documentation.
